refactor: 将 half-float 类型转换改为 CUDA 原生 intrinsic
使用 __half2float 和 __float2half_rn 替代 C++ 构造函数风格的
float(x) 和 __half(x) 转换，语义更清晰且符合 CUDA 规范。

+++ src/elementwise_gelu.cu
@@ -34,8 +34,8 @@ namespace elementwise_gelu
 
     __device__ __forceinline__ __half approximate_gelu_fast_tanh(__half x)
     {
-        float x_f = float(x);
-        return __half(x_f * 0.5f * (1.0f + fast_tanh(0.797884f * (x_f + 0.044715f * x_f * x_f * x_f))));
+        float x_f = __half2float(x);
+        return __float2half_rn(x_f * 0.5f * (1.0f + fast_tanh(0.797884f * (x_f + 0.044715f * x_f * x_f * x_f))));
     }
 
     __global__ void v0(__half* input, __half* output, const int size)
@@ -104,8 +104,8 @@ namespace elementwise_gelu
 
     __device__ __forceinline__ __half approximate_gelu(__half x)
     {
-        float x_f = float(x);
-        return __half(x_f * 0.5f * (1.0f + tanhf(0.797884f * (x_f + 0.044715f * x_f * x_f * x_f))));
+        float x_f = __half2float(x);
+        return __float2half_rn(x_f * 0.5f * (1.0f + tanhf(0.797884f * (x_f + 0.044715f * x_f * x_f * x_f))));
     }
 
     __global__ void v_ref(__half* input, __half* output, const int size)
